物品包枚举改成byte存储

+++ 模组代码/物品包/Items/类型_物品包.cs
@@ -12,7 +12,7 @@ namespace 物品包.Items;
 
 
 
-public enum 枚举_物品包类型 : int {
+public enum 枚举_物品包类型 : byte {
     物品包,
     嵌套包,
     饰品包,
